[refactor](renamefile) rename some files according to the class names
