+++ be/src/olap/merger.cpp
@@ -35,7 +35,7 @@
 #include "common/logging.h"
 #include "olap/olap_common.h"
 #include "olap/olap_define.h"
-#include "olap/reader.h"
+#include "olap/tablet_reader.h"
 #include "olap/rowid_conversion.h"
 #include "olap/rowset/rowset.h"
 #include "olap/rowset/rowset_meta.h"
