+++ be/src/olap/task/engine_checksum_task.cpp
@@ -27,7 +27,7 @@
 #include "io/io_common.h"
 #include "olap/olap_common.h"
 #include "olap/olap_define.h"
-#include "olap/reader.h"
+#include "olap/tablet_reader.h"
 #include "olap/rowset/rowset.h"
 #include "olap/storage_engine.h"
 #include "olap/tablet.h"
