+++ be/src/vec/sink/writer/vtablet_writer.cpp
@@ -53,7 +53,7 @@
 #include "vec/exprs/vexpr_fwd.h"
 #include "vec/runtime/vdatetime_value.h"
 #include "vec/sink/vrow_distribution.h"
-#include "vec/sink/vtablet_sink.h"
+#include "vec/sink/volap_table_sink.h"
 
 #ifdef DEBUG
 #include <unordered_set>
